nixpkgs-broken: add --list-pkg-paths command
This command can be used to list all package paths and add them to the database.
In the future this could be used to find packages where multiple attrs point to the same file.
In those cases, care should be taken to not just set a broken mark in that package.